+++ src/bin/main.rs
@@ -12,6 +12,7 @@ use console::{style, Term};
 use futures::{future::Either, Future, FutureExt};
 use indicatif::{MultiProgress, ProgressBar};
 use std::io::Write;
+use cli_clipboard::{ClipboardContext, ClipboardProvider};
 
 use magic_wormhole::{forwarding, transfer, transit, Wormhole};
 use std::str::FromStr;
@@ -288,6 +289,12 @@ async fn main() -> eyre::Result<()> {
             .try_init()?;
     }
 
+    let mut clipboard = ClipboardContext::new()
+        .map_err(|err| {
+            log::warn!("Failed to initialize clipboard support: {}", err);
+        })
+        .ok();
+
     let file_name = |file_path| {
         // TODO this has gotten out of hand (it ugly)
         // The correct solution would be to make `file_name` an Option everywhere and
@@ -331,7 +338,7 @@ async fn main() -> eyre::Result<()> {
                 matches,
                 true,
                 transfer::APP_CONFIG,
-                Some(&sender_print_code),
+                Some(&sender_print_code), clipboard.as_mut(),
             ),
             ctrl_c(),
         )
@@ -356,7 +363,7 @@ async fn main() -> eyre::Result<()> {
                 matches,
                 true,
                 transfer::APP_CONFIG,
-                Some(&sender_print_code),
+                Some(&sender_print_code), clipboard.as_mut(),
             );
             futures::pin_mut!(connect_fut);
             match futures::future::select(connect_fut, ctrl_c()).await {
@@ -388,7 +395,7 @@ async fn main() -> eyre::Result<()> {
 
         let (wormhole, _code, relay_server) = {
             let connect_fut =
-                parse_and_connect(&mut term, matches, false, transfer::APP_CONFIG, None);
+                parse_and_connect(&mut term, matches, false, transfer::APP_CONFIG, None, clipboard.as_mut());
             futures::pin_mut!(connect_fut);
             match futures::future::select(connect_fut, ctrl_c()).await {
                 Either::Left((result, _)) => result?,
@@ -453,7 +460,7 @@ async fn main() -> eyre::Result<()> {
                     matches,
                     true,
                     forwarding::APP_CONFIG,
-                    Some(&server_print_code),
+                    Some(&server_print_code), clipboard.as_mut(),
                 );
                 futures::pin_mut!(connect_fut);
                 let (wormhole, _code, relay_server) =
@@ -479,7 +486,7 @@ async fn main() -> eyre::Result<()> {
             let noconfirm = matches.is_present("noconfirm");
             let bind_address: std::net::IpAddr = matches.value_of("bind").unwrap().parse()?;
             let (wormhole, _code, relay_server) =
-                parse_and_connect(&mut term, matches, false, forwarding::APP_CONFIG, None).await?;
+                parse_and_connect(&mut term, matches, false, forwarding::APP_CONFIG, None, clipboard.as_mut()).await?;
             let relay_server = vec![transit::RelayHint::from_urls(None, [relay_server])];
 
             let offer =
@@ -521,6 +528,7 @@ async fn parse_and_connect(
     is_send: bool,
     mut app_config: magic_wormhole::AppConfig<impl serde::Serialize>,
     print_code: Option<&dyn Fn(&mut Term, &magic_wormhole::Code) -> eyre::Result<()>>,
+    clipboard: Option<&mut ClipboardContext>,
 ) -> eyre::Result<(Wormhole, magic_wormhole::Code, url::Url)> {
     let relay_server: url::Url = matches
         .value_of("relay-server")
@@ -561,7 +569,15 @@ async fn parse_and_connect(
             let (server_welcome, connector) =
                 magic_wormhole::Wormhole::connect_without_code(app_config, numwords).await?;
             print_welcome(term, &server_welcome)?;
+            /* Print code and also copy it to clipboard */
             if is_send {
+                if let Some(clipboard) = clipboard {
+                    match clipboard.set_contents(server_welcome.code.to_string()) {
+                        Ok(()) => log::info!("Code copied to clipboard"),
+                        Err(err) => log::warn!("Failed to copy code to clipboard: {}", err),
+                    }
+                }
+
                 print_code.expect("`print_code` must be `Some` when `is_send` is `true`")(
                     term,
                     &server_welcome.code,